+++ lib/generators/cms/component/profile_page/profile_page_generator.rb
@@ -1,31 +1,27 @@
-require 'generators/cms/migration'
-
 module Cms
   module Generators
     module Component
       class ProfilePageGenerator < ::Rails::Generators::Base
         include Migration
+        include BasePaths
 
         class_option :homepage_path,
           :type => :string,
-          :default => '/website/de',
+          :default => nil,
           :desc => 'Path to a CMS homepage, for which to create the contact form.'
 
         source_root File.expand_path('../templates', __FILE__)
 
-        def copy_app_directory
-          directory('app')
-          directory('config')
-          directory('spec')
-        end
-
-        def add_email_validation
+        def add_gems
           gem('valid_email', '0.0.4')
+          gem('localized_country_select', '>= 0.9.2')
+
+          Bundler.with_clean_env do
+            run('bundle --quiet')
+          end
         end
 
-        def add_country_select
-          gem('localized_country_select', '>= 0.9.2')
-          run('bundle install')
+        def import_translations
           run('rake import:country_select LOCALE=de')
           run('rake import:country_select LOCALE=en')
         end
@@ -76,39 +72,61 @@ module Cms
         end
 
         def create_migration
-          validate_obj_class(class_name)
+          begin
+            validate_attribute(profile_page_attribute_name)
 
-          migration_template('migration.rb', 'cms/migrate/create_profile_page.rb')
+            Rails::Generators.invoke('cms:attribute', [profile_page_attribute_name, '--type=linklist', '--title=Profile Page', '--max-size=1'])
+          rescue Cms::Generators::DuplicateResourceError
+          end
 
+          begin
+            validate_attribute(show_in_navigation_attribute_name)
+
+            Rails::Generators.invoke('cms:attribute', [show_in_navigation_attribute_name, '--type=boolean', '--title=Show in Navigation'])
+          rescue Cms::Generators::DuplicateResourceError
+          end
+
+          begin
+            validate_obj_class(class_name)
+
+            Rails::Generators.invoke('cms:model', [class_name, "--attributes=#{show_in_navigation_attribute_name}", '--title=Page: Profile'])
+          rescue Cms::Generators::DuplicateResourceError
+          end
+
+          migration_template('example_migration.rb', 'cms/migrate/create_profile_page_example.rb')
+        end
+
+        def copy_app_directory
+          directory('app', force: true)
+          directory('config', force: true)
+          directory('spec', force: true)
+        end
+
+        def notice
           if behavior == :invoke
             log(:migration, 'Make sure to run "rake cms:migrate" to apply CMS changes and "bundle" to install new gem.')
           end
-        rescue DuplicateResourceError
-        end
-
-        def cms_migrate
-          run('rake cms:migrate')
         end
 
         private
 
+        alias_method :original_homepage_path, :homepage_path
         def homepage_path
-          options[:homepage_path]
+          options[:homepage_path] || original_homepage_path
+        end
+
+        def show_in_navigation_attribute_name
+          'show_in_navigation'
+        end
+
+        def profile_page_attribute_name
+          'profile_page_link'
         end
 
         def class_name
           'ProfilePage'
         end
-
       end
     end
   end
 end
-#
-#        def create_structure_migration_file
-#          Rails::Generators.invoke('cms:attribute', ['profile_page', '--type=linklist'])
-#          Rails::Generators.invoke('cms:model', ['ProfilePage', '--title=Page: Profile', '--attributes=show_in_navigation'])
-#
-#          migration_template('create_profilepage_structure.rb', 'cms/migrate/create_profilepage_structure.rb')
-#        end
-