+++ lib/generators/cms/component/profile_page/templates/app/concerns/cms/attributes/profile_page_link.rb
@@ -1,18 +1,8 @@
 module Cms
   module Attributes
     module ProfilePageLink
-      def profile_page_link_attribute
-        :profile_page_link
-      end
-
-      def default_profile_page_link
-        RailsConnector::LinkList.new(nil)
-      ensure
-        Cms::Errors::MissingAttribute.notify(self, __FILE__)
-      end
-
       def profile_page_link
-        self[profile_page_link_attribute] || default_profile_page_link
+        self[:profile_page_link] || RailsConnector::LinkList.new(nil)
       end
 
       def profile_page_link?