+++ lib/generators/cms/component/profile_page/templates/app/controllers/profile_page_controller.rb
@@ -4,9 +4,9 @@ class ProfilePageController < CmsController
   before_filter Filters::Authorization
 
   def index
-    @profile_page_presenter = ProfilePagePresenter.new(current_user, params[:profile_page_presenter])
+    @presenter = ProfilePagePresenter.new(current_user, params[:profile_page_presenter])
 
-    if request.post? && @profile_page_presenter.save
+    if request.post? && @presenter.save
       flash.now[:notice] = I18n.t('flash.profile_page.success')
     end
   end