+++ lib/generators/cms/component/profile_page/templates/app/models/profile_page.rb
@@ -1,4 +1,0 @@
-class ProfilePage < Obj
-  include Cms::Attributes::ShowInNavigation
-  include Page
-end