+++ lib/ice_kickstarter/rake/integration_task.rb
@@ -53,6 +53,7 @@ module IceKickstarter
           sh "cd #{app_path} && bundle exec rails generate cms:component:google_analytics"
           sh "cd #{app_path} && bundle exec rails generate cms:component:contact_page"
           sh "cd #{app_path} && bundle exec rails generate cms:component:language_switch"
+          sh "cd #{app_path} && bundle exec rails generate cms:component:profile_page"
         end
       end